+++ NEWS.md
@@ -1,3 +1,8 @@
+# cccapi 0.1.2 (2022-03-22)
+
+* Safer queries in `get_search()` *via* `utils::URLencode()`.
+* Minor (non-functional) code cosmetics.
+
 # cccapi 0.1.1 (2022-01-26)
 
 * Fixed a bug when searching for SMILES with URL-problematic characters.